+++ src/components/Navigation/Burger.js
@@ -1,4 +1,26 @@
-import styled from 'styled-components';
+import React from 'react';
+import { bool, func } from 'prop-types';
+import styled from 'styled-components'
+
+const Burger = ({ open, setOpen, ...props }) => {
+  
+  const isExpanded = open ? true : false;
+  
+  return (
+    <StyledBurger aria-label="Toggle menu" aria-expanded={isExpanded} open={open} onClick={() => setOpen(!open)} {...props}>
+      <span />
+      <span />
+      <span />
+    </StyledBurger>
+  )
+}
+
+Burger.propTypes = {
+  open: bool.isRequired,
+  setOpen: func.isRequired,
+};
+
+export default Burger;
 
 export const StyledBurger = styled.button`
   position: absolute;
@@ -38,4 +60,4 @@ export const StyledBurger = styled.button`
       transform: ${({ open }) => open ? 'rotate(-45deg)' : 'rotate(0)'};
     }
   }
-`;
+`
