--- src/components/Navigation/Burger/index.js
@@ -1,1 +1,0 @@
-export { default } from './Burger';
